Make warning-free in 2015 edition.
The std::error::Error trait has deprecated the `description` method in favor of
an implementation of the `Display` trait, and `cause` is deprecated in favor of
`source`, which adds a `'static` bound on the return type so downcasting is
permitted.

The `try!` macro is deprecated in favor of the `?` operator.

Use the `dyn` keyword in trait objects.